requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0

sentencepiece>=0.1.99
protobuf>=3.20.0
cpm-kernels>=1.0.11

numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
datasets>=2.14.0

sentence-transformers>=2.2.0
scipy>=1.11.0

neo4j>=5.12.0
py2neo>=2021.2.3

matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

jieba>=0.42.1
opencc-python-reimplemented>=0.1.7

pyyaml>=6.0
omegaconf>=2.3.0

ipython>=8.14.0
jupyter>=1.0.0
tensorboard>=2.13.0

requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

colorlog>=6.7.0
wandb>=0.15.0
